requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.1.4
reportlab==4.0.7
opencv-python==4.8.1.78
Pillow==10.1.0